extracted decoders to decode file and updated dist
