+++ README.md
@@ -29,7 +29,12 @@ console.log(encoded) // UDfpLD2rKt2oKl18pSd91m0KGWDz
 
 var url = 'http://www.plantuml.com/plantuml/img/' + encoded
 ```
-# decode example
+
+The URL can then be used to display the diagram:
+
+![alt tag](http://www.plantuml.com/plantuml/img/UDfpLD2rKt2oKl18pSd91m0KGWDz)
+
+# Decode example
 
 ```javascript
 var plantumlEncoder = require('plantuml-encoder')
@@ -38,10 +43,6 @@ var plain = plantumlEncoder.decode('UDfpLD2rKt2oKl18pSd91m0KGWDz')
 console.log(plain) // A -> B: Hello
 
 ```
-
-The URL can then be used to display the diagram:
-
-![alt tag](http://www.plantuml.com/plantuml/img/UDfpLD2rKt2oKl18pSd91m0KGWDz)
 
 # License
 MIT
